Revert "Ensemble copula coupling extension"
