+++ lib/improver/constants.py
@@ -33,9 +33,6 @@
 # Real Missing Data Indicator
 RMDI = -32767.0
 
-# 0 Kelvin in degrees C
-ABSOLUTE_ZERO = 273.15
-
 # Specific gas constant for dry air (J K-1 kg-1)
 R_DRY_AIR = 287.0
