+++ lib/improver/tests/helper_functions_ensemble_calibration.py
@@ -35,7 +35,7 @@ Functions for use within unit tests for `ensemble_calibration` plugins.
 
 from cf_units import Unit
 import iris
-from iris.coords import AuxCoord, DimCoord
+from iris.coords import DimCoord
 from iris.cube import Cube, CubeList
 import numpy as np
 
@@ -43,117 +43,21 @@ from improver.ensemble_calibration.ensemble_calibration_utilities import (
     concatenate_cubes)
 
 
-def set_up_probability_above_threshold_cube(
-        data, phenomenon_standard_name, phenomenon_units,
-        forecast_thresholds=np.array([8, 10, 12]), timesteps=1,
-        y_dimension_length=3, x_dimension_length=3):
-    """
-    Create a cube containing multiple probability_above_threshold
-    values for the coordinate.
-    """
-    cube_long_name = (
-        "probability_of_{}_above_threshold".format(phenomenon_standard_name))
-    cube = Cube(data, long_name=cube_long_name,
+def set_up_cube(data, phenomenon_standard_name, phenomenon_units):
+    """Create a cube containing multiple realizations."""
+    cube = Cube(data, standard_name=phenomenon_standard_name,
                 units=phenomenon_units)
-    coord_long_name = "{}_threshold".format(phenomenon_standard_name)
-    cube.add_dim_coord(
-        DimCoord(forecast_thresholds, long_name=coord_long_name,
-                 units='degreesC'), 0)
-    time_origin = "hours since 1970-01-01 00:00:00"
-    calendar = "gregorian"
-    tunit = Unit(time_origin, calendar)
-    cube.add_dim_coord(DimCoord(np.linspace(402192.5, 402292.5, timesteps),
-                                "time", units=tunit), 1)
-    cube.add_dim_coord(DimCoord(np.linspace(-45.0, 45.0, y_dimension_length),
-                                'latitude', units='degrees'), 2)
-    cube.add_dim_coord(DimCoord(np.linspace(120, 180, x_dimension_length),
-                                'longitude', units='degrees'), 3)
-    return cube
-
-
-def set_up_probability_above_threshold_temperature_cube():
-    """
-    Create a cube with metadata and values suitable for air temperature.
-    """
-    data = np.array([[[[1.0, 0.9, 1.0],
-                       [0.8, 0.9, 0.5],
-                       [0.5, 0.2, 0.0]]],
-                     [[[1.0, 0.5, 1.0],
-                       [0.5, 0.5, 0.3],
-                       [0.2, 0.0, 0.0]]],
-                     [[[1.0, 0.2, 0.5],
-                       [0.2, 0.0, 0.1],
-                       [0.0, 0.0, 0.0]]]])
-    return (
-        set_up_probability_above_threshold_cube(data, "air_temperature", "1"))
-
-
-def set_up_probability_above_threshold_spot_cube(
-        data, phenomenon_standard_name, phenomenon_units,
-        forecast_thresholds=np.array([8, 10, 12]),
-        y_dimension_length=9, x_dimension_length=9):
-    """
-    Create a cube containing multiple realizations, where one of the
-    dimensions is an index used for spot forecasts.
-    """
-    cube_long_name = (
-        "probability_of_{}_above_threshold".format(phenomenon_standard_name))
-    cube = Cube(data, long_name=cube_long_name,
-                units=phenomenon_units)
-    coord_long_name = "{}_threshold".format(phenomenon_standard_name)
-    cube.add_dim_coord(
-        DimCoord(forecast_thresholds, long_name=coord_long_name,
-                 units='degreesC'), 0)
+    cube.add_dim_coord(DimCoord([0, 1, 2], 'realization',
+                                units='1'), 0)
     time_origin = "hours since 1970-01-01 00:00:00"
     calendar = "gregorian"
     tunit = Unit(time_origin, calendar)
     cube.add_dim_coord(DimCoord([402192.5],
                                 "time", units=tunit), 1)
-    cube.add_dim_coord(DimCoord(np.arange(9), long_name='locnum',
-                                units="1"), 2)
-    cube.add_aux_coord(AuxCoord(np.linspace(-45.0, 45.0, y_dimension_length),
-                                'latitude', units='degrees'), data_dims=2)
-    cube.add_aux_coord(AuxCoord(np.linspace(120, 180, x_dimension_length),
-                                'longitude', units='degrees'), data_dims=2)
-    return cube
-
-
-def set_up_probability_above_threshold_spot_temperature_cube():
-    """
-    Create a cube with metadata and values suitable for air temperature
-    for spot forecasts.
-    """
-    data = np.array([[[1.0, 0.9, 1.0,
-                       0.8, 0.9, 0.5,
-                       0.5, 0.2, 0.0]],
-                     [[1.0, 0.5, 1.0,
-                       0.5, 0.5, 0.3,
-                       0.2, 0.0, 0.0]],
-                     [[1.0, 0.2, 0.5,
-                       0.2, 0.0, 0.1,
-                       0.0, 0.0, 0.0]]])
-    return (
-        set_up_probability_above_threshold_spot_cube(
-            data, "air_temperature", "1"))
-
-
-def set_up_cube(data, phenomenon_standard_name, phenomenon_units,
-                realizations=np.array([0, 1, 2]), timesteps=1,
-                y_dimension_length=3, x_dimension_length=3):
-    """Create a cube containing multiple realizations."""
-    cube = Cube(data, standard_name=phenomenon_standard_name,
-                units=phenomenon_units)
-    cube.add_dim_coord(DimCoord(realizations, 'realization',
-                                units='1'), 0)
-    time_origin = "hours since 1970-01-01 00:00:00"
-    calendar = "gregorian"
-    tunit = Unit(time_origin, calendar)
-    cube.add_dim_coord(DimCoord(np.linspace(402192.5, 402292.5, timesteps),
-                                "time", units=tunit), 1)
-    cube.add_dim_coord(DimCoord(np.linspace(-45.0, 45.0, y_dimension_length),
-                                'latitude', units='degrees'), 2)
-    cube.add_dim_coord(DimCoord(np.linspace(120, 180, x_dimension_length),
-                                'longitude', units='degrees'), 3)
+    cube.add_dim_coord(DimCoord(np.linspace(-45.0, 45.0, 3), 'latitude',
+                                units='degrees'), 2)
+    cube.add_dim_coord(DimCoord(np.linspace(120, 180, 3), 'longitude',
+                                units='degrees'), 3)
     return cube
 
 
@@ -165,36 +69,6 @@ def set_up_temperature_cube():
     data[1] += 2
     data[2] += 4
     return set_up_cube(data, "air_temperature", "K")
-
-
-def set_up_spot_cube(data, phenomenon_standard_name, phenomenon_units):
-    """Create a cube containing multiple realizations."""
-    cube = Cube(data, standard_name=phenomenon_standard_name,
-                units=phenomenon_units)
-    cube.add_dim_coord(DimCoord([0, 1, 2], 'realization',
-                                units='1'), 0)
-    time_origin = "hours since 1970-01-01 00:00:00"
-    calendar = "gregorian"
-    tunit = Unit(time_origin, calendar)
-    cube.add_dim_coord(DimCoord([402192.5],
-                                "time", units=tunit), 1)
-    cube.add_dim_coord(DimCoord(np.arange(9), long_name='locnum',
-                                units="1"), 2)
-    cube.add_aux_coord(AuxCoord(np.linspace(-45.0, 45.0, 9), 'latitude',
-                                units='degrees'), data_dims=2)
-    cube.add_aux_coord(AuxCoord(np.linspace(120, 180, 9), 'longitude',
-                                units='degrees'), data_dims=2)
-    return cube
-
-
-def set_up_spot_temperature_cube():
-    """Create a cube with metadata and values suitable for air temperature."""
-    data = (np.tile(np.linspace(-45.0, 45.0, 9), 3).reshape(3, 1, 9) +
-            273.15)
-    data[0] -= 2
-    data[1] += 2
-    data[2] += 4
-    return set_up_spot_cube(data, "air_temperature", "K")
 
 
 def set_up_wind_speed_cube():
