+++ lib/improver/tests/test_ensemble_calibration_EnsembleCalibrationUtilities.py
@@ -44,8 +44,8 @@ from iris.tests import IrisTest
 import numpy as np
 
 from improver.ensemble_calibration.ensemble_calibration_utilities import (
-    convert_cube_data_to_2d, ensure_dimension_is_the_zeroth_dimension,
-    concatenate_cubes, _associate_any_coordinate_with_master_coordinate,
+    convert_cube_data_to_2d, concatenate_cubes,
+    _associate_any_coordinate_with_master_coordinate,
     _slice_over_coordinate, _strip_var_names, rename_coordinate, _renamer,
     check_predictor_of_mean_flag)
 from improver.tests.helper_functions_ensemble_calibration import(
@@ -195,69 +195,6 @@ class Test_convert_cube_data_to_2d(IrisTest):
 
         result = convert_cube_data_to_2d(cube)
         self.assertArrayAlmostEqual(result, data)
-
-
-class Test_ensure_dimension_is_the_zeroth_dimension(IrisTest):
-
-    """
-    Test the ensure_dimension_is_the_zeroth_dimension
-    utility.
-    """
-
-    def setUp(self):
-        """Use temperature cube to test with."""
-        self.cube = set_up_temperature_cube()
-
-    def test_basic(self):
-        """Test that the function returns an iris.cube.Cube."""
-        result = (
-            ensure_dimension_is_the_zeroth_dimension(self.cube, "realization"))
-        self.assertIsInstance(result, Cube)
-
-    def test_if_probabilistic_dimension_is_first(self):
-        """
-        Test that a cube with the expected data contents is returned when
-        the probabilistic dimension is the first dimension coordinate.
-        """
-        result = (
-            ensure_dimension_is_the_zeroth_dimension(self.cube, "realization"))
-        self.assertArrayAlmostEqual(result.data, self.cube.data)
-
-    def test_if_probabilistic_dimension_is_not_first(self):
-        """
-        Test that a cube with the expected data contents is returned when
-        the probabilistic dimension is a dimension coordinate but it is
-        not the first dimension coordinate,.
-        """
-        expected = self.cube.copy()
-        expected.transpose([0, 3, 2, 1])
-
-        cube = self.cube.copy()
-        cube.transpose([3, 2, 1, 0])
-        result = (
-            ensure_dimension_is_the_zeroth_dimension(cube, "realization"))
-        self.assertArrayAlmostEqual(result.data, expected.data)
-
-    def test_if_probabilistic_dimension_is_scalar(self):
-        """
-        Test that a cube with the expected data contents is returned when
-        the probabilistic dimension is a scalar coordinate.
-        """
-        cube = self.cube[0, :, :, :]
-        result = (
-            ensure_dimension_is_the_zeroth_dimension(cube, "realization"))
-        self.assertArrayAlmostEqual(result.data, [cube.data])
-
-    def test_if_probabilistic_dimension_not_available(self):
-        """
-        Test that the expected error message is raised when the required
-        probabilistic dimension is not available in the cube.
-        """
-        cube = self.cube[0, :, :, :]
-        cube.remove_coord("realization")
-        msg = "not a dimension coordinate"
-        with self.assertRaisesRegexp(ValueError, msg):
-            ensure_dimension_is_the_zeroth_dimension(cube, "realization")
 
 
 class Test_concatenate_cubes(IrisTest):
