+++ lib/improver/tests/test_ensemble_calibration_GeneratePercentilesFromMeanAndVariance.py
@@ -30,21 +30,91 @@
 # POSSIBILITY OF SUCH DAMAGE.
 """
 Unit tests for the
-`ensemble_copula_coupling.GeneratePercentilesFromMeanAndVariance`
+`ensemble_calibration.GeneratePercentilesFromMeanAndVariance`
+class.
 
 """
 import unittest
 
 import iris
+from iris.coords import DimCoord
 from iris.cube import Cube, CubeList
 from iris.tests import IrisTest
 import numpy as np
 
-from improver.ensemble_copula_coupling.ensemble_copula_coupling import (
+from improver.ensemble_calibration.ensemble_calibration import (
     GeneratePercentilesFromMeanAndVariance as Plugin)
 from improver.tests.helper_functions_ensemble_calibration import(
-    set_up_spot_temperature_cube, set_up_temperature_cube,
-    add_forecast_reference_time_and_forecast_period)
+    set_up_temperature_cube, add_forecast_reference_time_and_forecast_period)
+
+
+class Test__create_cube_with_percentiles(IrisTest):
+
+    """Test the _create_cube_with_percentiles plugin."""
+
+    def setUp(self):
+        """Set up temperature cube."""
+        self.current_temperature_forecast_cube = (
+            add_forecast_reference_time_and_forecast_period(
+                set_up_temperature_cube()))
+
+    def test_basic(self):
+        """Test that the plugin returns an Iris.cube.Cube."""
+        cube = self.current_temperature_forecast_cube
+        cube_data = cube.data + 2
+        percentiles = [0.1, 0.5, 0.9]
+        plugin = Plugin()
+        result = plugin._create_cube_with_percentiles(
+            percentiles, cube, cube_data)
+        self.assertIsInstance(result, Cube)
+
+    def test_many_percentiles(self):
+        """
+        Test that the plugin returns an Iris.cube.Cube with many
+        percentiles.
+        """
+        cube = self.current_temperature_forecast_cube
+        percentiles = np.linspace(0, 1, 100)
+        cube_data = np.zeros(
+            [len(percentiles), len(cube.coord("time").points),
+             len(cube.coord("latitude").points),
+             len(cube.coord("longitude").points)])
+        plugin = Plugin()
+        result = plugin._create_cube_with_percentiles(
+            percentiles, cube, cube_data)
+        self.assertEqual(cube_data.shape, result.data.shape)
+
+    def test_incompatible_percentiles(self):
+        """
+        Test that the plugin fails if the percentile values requested
+        are not numbers.
+        """
+        cube = self.current_temperature_forecast_cube
+        percentiles = ["cat", "dog", "elephant"]
+        cube_data = np.zeros(
+            [len(percentiles), len(cube.coord("time").points),
+             len(cube.coord("latitude").points),
+             len(cube.coord("longitude").points)])
+        plugin = Plugin()
+        msg = "could not convert string to float"
+        with self.assertRaisesRegexp(ValueError, msg):
+            plugin._create_cube_with_percentiles(
+                percentiles, cube, cube_data)
+
+    def test_percentile_points(self):
+        """
+        Test that the plugin returns an Iris.cube.Cube
+        with a percentile coordinate with the desired points.
+        """
+        cube = self.current_temperature_forecast_cube
+        cube_data = cube.data + 2
+        percentiles = [0.1, 0.5, 0.9]
+        plugin = Plugin()
+        result = plugin._create_cube_with_percentiles(
+            percentiles, cube, cube_data)
+        self.assertIsInstance(result.coord("percentile"), DimCoord)
+        self.assertArrayAlmostEqual(
+            result.coord("percentile").points, percentiles)
 
 
 class Test__mean_and_variance_to_percentiles(IrisTest):
@@ -56,9 +126,6 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
         self.current_temperature_forecast_cube = (
             add_forecast_reference_time_and_forecast_period(
                 set_up_temperature_cube()))
-        self.current_temperature_spot_forecast_cube = (
-            add_forecast_reference_time_and_forecast_period(
-                set_up_spot_temperature_cube()))
 
     def test_check_data(self):
         """
@@ -82,6 +149,8 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
             "realization", iris.analysis.MEAN)
         current_forecast_variance = cube.collapsed(
             "realization", iris.analysis.VARIANCE)
+        current_forecast_predictor_and_variance = (
+            current_forecast_predictor, current_forecast_variance)
         percentiles = [0.1, 0.5, 0.9]
         plugin = Plugin()
         result = plugin._mean_and_variance_to_percentiles(
@@ -122,6 +191,8 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
             "realization", iris.analysis.MEAN)
         current_forecast_variance = cube.collapsed(
             "realization", iris.analysis.VARIANCE)
+        current_forecast_predictor_and_variance = (
+            current_forecast_predictor, current_forecast_variance)
         percentiles = [0.1, 0.5, 0.9]
         plugin = Plugin()
         result = plugin._mean_and_variance_to_percentiles(
@@ -159,6 +230,8 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
             "realization", iris.analysis.MEAN)
         current_forecast_variance = cube.collapsed(
             "realization", iris.analysis.VARIANCE)
+        current_forecast_predictor_and_variance = (
+            current_forecast_predictor, current_forecast_variance)
         percentiles = [0.1, 0.5, 0.9]
         plugin = Plugin()
         result = plugin._mean_and_variance_to_percentiles(
@@ -185,7 +258,7 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
                            [3., 3., 3.]]]])
 
         result_data = np.array([[[[1., 1., 1.],
-                                  [1.18685838, 2., 2.],
+                                  [1.186858, 2., 2.],
                                   [3., 3., 3.]]],
                                 [[[1., 1., 1.],
                                   [2.66666667, 2., 2.],
@@ -200,6 +273,8 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
             "realization", iris.analysis.MEAN)
         current_forecast_variance = cube.collapsed(
             "realization", iris.analysis.VARIANCE)
+        current_forecast_predictor_and_variance = (
+            current_forecast_predictor, current_forecast_variance)
         percentiles = [0.1, 0.5, 0.9]
         plugin = Plugin()
         result = plugin._mean_and_variance_to_percentiles(
@@ -217,6 +292,8 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
             "realization", iris.analysis.MEAN)
         current_forecast_variance = cube.collapsed(
             "realization", iris.analysis.VARIANCE)
+        current_forecast_predictor_and_variance = (
+            current_forecast_predictor, current_forecast_variance)
         percentiles = np.linspace(0.01, 0.99, num=1000, endpoint=True)
         plugin = Plugin()
         result = plugin._mean_and_variance_to_percentiles(
@@ -233,43 +310,78 @@ class Test__mean_and_variance_to_percentiles(IrisTest):
             "realization", iris.analysis.MEAN)
         current_forecast_variance = cube.collapsed(
             "realization", iris.analysis.VARIANCE)
+        current_forecast_predictor_and_variance = (
+            current_forecast_predictor, current_forecast_variance)
         percentiles = [-0.1, 0.1]
         plugin = Plugin()
         msg = "NaNs are present within the result for the"
         with self.assertRaisesRegexp(ValueError, msg):
-            plugin._mean_and_variance_to_percentiles(
+            result = plugin._mean_and_variance_to_percentiles(
                 current_forecast_predictor, current_forecast_variance,
                 percentiles)
 
-    def test_spot_forecasts_check_data(self):
-        """
-        Test that the plugin returns an Iris.cube.Cube matching the expected
-        data values when a cube containing mean and variance is passed in.
-        The resulting data values are the percentiles, which have been
-        generated for a spot forecast.
-        """
-        data = np.array([[[225.56812863, 236.81812863, 248.06812863,
-                           259.31812863, 270.56812863, 281.81812863,
-                           293.06812863, 304.31812863, 315.56812863]],
-                         [[229.48333333, 240.73333333, 251.98333333,
-                           263.23333333, 274.48333333, 285.73333333,
-                           296.98333333, 308.23333333, 319.48333333]],
-                         [[233.39853804, 244.64853804, 255.89853804,
-                           267.14853804, 278.39853804, 289.64853804,
-                           300.89853804, 312.14853804, 323.39853804]]])
 
-        cube = self.current_temperature_spot_forecast_cube
-        current_forecast_predictor = cube.collapsed(
-            "realization", iris.analysis.MEAN)
-        current_forecast_variance = cube.collapsed(
-            "realization", iris.analysis.VARIANCE)
-        percentiles = [0.1, 0.5, 0.9]
+class Test_create_percentiles(IrisTest):
+
+    """Test the create_percentiles plugin."""
+
+    def setUp(self):
+        """Set up temperature cube."""
+        self.current_temperature_forecast_cube = (
+            add_forecast_reference_time_and_forecast_period(
+                set_up_temperature_cube()))
+
+    def test_basic(self):
+        """
+        Test that the plugin returns a list with the expected number of
+        percentiles.
+        """
+        cube = self.current_temperature_forecast_cube
+        no_of_percentiles = 3
         plugin = Plugin()
-        result = plugin._mean_and_variance_to_percentiles(
-            current_forecast_predictor, current_forecast_variance,
-            percentiles)
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(result.data, data)
+        result = plugin._create_percentiles(no_of_percentiles)
+        self.assertIsInstance(result, list)
+        self.assertEqual(len(result), no_of_percentiles)
+
+    def test_data(self):
+        """
+        Test that the plugin returns a list with the expected data values
+        for the percentiles.
+        """
+        data = np.array([0.25, 0.5, 0.75])
+
+        cube = self.current_temperature_forecast_cube
+        no_of_percentiles = 3
+        plugin = Plugin()
+        result = plugin._create_percentiles(no_of_percentiles)
+        self.assertArrayAlmostEqual(result, data)
+
+    def test_random(self):
+        """
+        Test that the plugin returns a list with the expected number of
+        percentiles, if the random sampling option is selected.
+        """
+        cube = self.current_temperature_forecast_cube
+        no_of_percentiles = 3
+        plugin = Plugin()
+        result = plugin._create_percentiles(
+            no_of_percentiles, sampling="random")
+        self.assertIsInstance(result, list)
+        self.assertEqual(len(result), no_of_percentiles)
+
+    def test_unknown_sampling_option(self):
+        """
+        Test that the plugin returns the expected error message,
+        if an unknown sampling option is selected.
+        """
+        cube = self.current_temperature_forecast_cube
+        no_of_percentiles = 3
+        plugin = Plugin()
+
+        msg = "The unknown sampling option is not yet implemented"
+        with self.assertRaisesRegexp(ValueError, msg):
+            plugin._create_percentiles(
+                no_of_percentiles, sampling="unknown")
 
 
 class Test_process(IrisTest):
@@ -293,10 +405,9 @@ class Test_process(IrisTest):
 
         predictor_and_variance = CubeList(
             [current_forecast_predictor, current_forecast_variance])
-        no_of_percentiles = len(raw_forecast.coord("realization").points)
 
         plugin = Plugin()
-        result = plugin.process(predictor_and_variance, no_of_percentiles)
+        result = plugin.process(predictor_and_variance, raw_forecast)
         self.assertIsInstance(result, Cube)
 
     def test_number_of_percentiles(self):
@@ -314,10 +425,8 @@ class Test_process(IrisTest):
         predictor_and_variance = CubeList(
             [current_forecast_predictor, current_forecast_variance])
 
-        no_of_percentiles = len(raw_forecast.coord("realization").points)
-
         plugin = Plugin()
-        result = plugin.process(predictor_and_variance, no_of_percentiles)
+        result = plugin.process(predictor_and_variance, raw_forecast)
         self.assertEqual(len(raw_forecast.coord("realization").points),
                          len(result.coord("percentile").points))
